+++ src/Data/TypedEncoding/Conv/ByteString/Lazy/Char8.hs
@@ -17,10 +17,20 @@ import           Data.TypedEncoding.Instances.Support
 
 -- | 
 -- Lazy version of 'Data.TypedEncoding.Conv.ByteString.Char8.pack'.
-pack :: (Knds.LLast xs ~ t, IsSuperset "r-ASCII" t ~ 'True) => Enc xs c String -> Enc xs c BL8.ByteString
+pack :: (
+     Knds.UnSnoc xs ~ '(,) ys y
+    , Superset "r-CHAR8" y 
+    , encs ~ RemoveRs ys
+    , AllEncodeInto "r-CHAR8" encs
+    ) => Enc xs c String -> Enc xs c BL8.ByteString
 pack = unsafeChangePayload BL8.pack
 
 -- | 
 -- Lazy version of 'Data.TypedEncoding.Conv.ByteString.Char8.unpack'.
-unpack :: (Knds.LLast xs ~ t, IsSuperset "r-ASCII" t ~ 'True) => Enc xs c BL8.ByteString -> Enc xs c String
+unpack :: (
+          Knds.UnSnoc xs ~ '(,) ys y
+        , Superset "r-CHAR8" y
+        , encs ~ RemoveRs ys
+        , AllEncodeInto "r-CHAR8" encs
+       ) => Enc xs c BL8.ByteString -> Enc xs c String
 unpack = unsafeChangePayload BL8.unpack